[BUGFIX] Fixed inconsistent type

+++ utils/entropy.h
@@ -20,7 +20,7 @@ namespace pesieve {
             //
             for (it = counts.begin(); it != counts.end(); ++it) {
                 double p_x = (double)it->second / elements;
-                if (p_x > 0) entropy -= p_x * log(p_x) / log(2);
+                if (p_x > 0) entropy -= p_x * log(p_x) / log((double)2);
             }
             return entropy;
         }
